Move CoreData models into their own files with extensions
