+++ EssentialFeed/EssentialFeed/Feed Cache/Infastructure/Core Data/CoreDataFeedStore.swift
@@ -70,50 +70,3 @@ public class CoreDataFeedStore: FeedStore {
         action(context)
     }
 }
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-    @NSManaged var timestamp: Date
-    @NSManaged var feed: NSOrderedSet
-    
-    var localFeed: [LocalFeedImage] {
-        return feed
-            .compactMap{ ($0 as? ManagedFeedImage)?.local }
-    }
-    
-    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
-        let request: NSFetchRequest<ManagedCache> = NSFetchRequest(entityName: ManagedCache.entity().name!)
-        request.returnsObjectsAsFaults = false
-        
-        return try context.fetch(request).first
-    }
-    
-    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
-        try find(in: context).map(context.delete)
-        return ManagedCache(context: context)
-    }
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-    @NSManaged var id: UUID
-    @NSManaged var imageDescription: String?
-    @NSManaged var location: String?
-    @NSManaged var url: URL
-    @NSManaged var cache: ManagedCache
-    
-    static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet {
-        return NSOrderedSet(array: localFeed.map { local in
-            let managedItem = ManagedFeedImage(context: context)
-            managedItem.id = local.id
-            managedItem.imageDescription = local.description
-            managedItem.location = local.location
-            managedItem.url = local.url
-            return managedItem
-        })
-    }
-    
-    var local: LocalFeedImage {
-        return  LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
-    }
-}
